pyproject.toml
[project]
name = "agent-identity"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pydantic>=2.12.5",
    "pylatex>=1.4.2",
    "pyyaml>=6.0.3",
    "qrcode[pil]>=8.2",
]

[dependency-groups]
dev = [
    "pip-audit>=2.10.0",
    "pre-commit>=4.5.1",
    "pyright>=1.1.408",
    "ruff>=0.15.2",
]

[tool.ruff]
exclude = ["typings", ".venv"]

[tool.ruff.lint]
select = ["E", "W", "F", "I", "TID"]

[tool.ruff.format]
docstring-code-format = true

[tool.pyright]
venvPath = "."
venv = ".venv"
exclude = ["typings", ".venv", ".vscode", ".git"]
